third_party/blink/web_tests/http/tests/cache/history-only-cached-subresource-loads-max-age-https-expected.txt

CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/cache/resources/max-age-resource.html' attempted to load subresource at 'https://127.0.0.1:8443/cache/resources/random-max-age.cgi' over broken HTTPS.
CONSOLE ERROR: line 1: Mixed Content: The page at 'https://127.0.0.1:8443/cache/resources/max-age-resource.html' attempted to load subresource at 'https://127.0.0.1:8443/cache/resources/random-max-age.cgi' over broken HTTPS.
CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/cache/resources/max-age-resource.html' attempted to load subresource at 'https://127.0.0.1:8443/cache/resources/random-max-age.cgi' over broken HTTPS.
CONSOLE ERROR: Mixed Content: The page at 'https://127.0.0.1:8443/cache/resources/max-age-resource-next.html' attempted to load subresource at 'https://127.0.0.1:8443/cache/resources/random-max-age.cgi' over broken HTTPS.
This test checks that loading a subresource with "Cache-Control: max-age=0" is cached and reused in back navigation when the page is not in the page cache.

We then test that loading the same subresource is refetched when used in non-stale loads such as refreshes or normal navigation.

PASS - max-age subresource was cached and used for a back navigation
PASS - max-age subresource was refetched with a reload
PASS - max-age subresource was refetched with a normal navigation
